# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.BanIPRequest:
    properties:
      expires_at:
        type: string
      ip:
        type: string
      reason:
        type: string
    type: object
...
  handler.ListUsersResponse:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
...
  handler.RejectCashPaymentRequest:
    properties:
      reason:
        type: string
    type: object
...
  handler.SetRegionSwitchPriceRequest:
    properties:
      price:
        type: number
    type: object
...
  model.Plan:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_days:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      max_devices:
        type: integer
      name:
        type: string
      price_stars:
        type: integer
      price_ton:
        type: number
      price_usd:
        type: number
      sort_order:
        type: integer
      traffic_gb:
        type: integer
      visible_to_referrer_id:
        description: |-
          VisibleToReferrerID restricts plan visibility to users referred by
          this user. NULL means public to everyone.
        type: integer
    type: object
...
host: api.zaruchevskiy.ru
info:
  contact: {}
  description: |-
    REST API of the ZyVPN Telegram Mini App backend. All
    /api/* routes require X-Telegram-Init-Data; /api/admin/*
    additionally requires admin role in the database.
  title: ZyVPN API
  version: "1.0"
paths:
  /api/admin/bans:
    get:
      parameters:
      - default: 50
        description: page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: List bans
      tags:
      - admin
  /api/admin/bans/ip:
    post:
      consumes:
      - application/json
      parameters:
      - description: ip + reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BanIPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Ban IP
      tags:
      - admin
  /api/admin/bans/ip/unban:
    post:
      consumes:
      - application/json
      parameters:
      - description: '{ip}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Unban IP
      tags:
      - admin
  /api/admin/logs:
    get:
      parameters:
      - default: 50
        description: page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Admin audit log
      tags:
      - admin
  /api/admin/payments/cash/{payment_id}/approve:
    post:
      parameters:
      - description: payment uuid
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Approve cash payment
      tags:
      - admin
  /api/admin/payments/cash/{payment_id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: payment uuid
        in: path
        name: payment_id
        required: true
        type: string
      - description: reason
        in: body
        name: body
        schema:
          $ref: '#/definitions/handler.RejectCashPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Reject cash payment
      tags:
      - admin
  /api/admin/payments/cash/pending:
    get:
      parameters:
      - default: 50
        description: page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Pending cash payments
      tags:
      - admin
  /api/admin/plans:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Plan'
            type: array
      security:
      - TelegramInitData: []
      summary: List all plans
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: new plan
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Plan'
      security:
      - TelegramInitData: []
      summary: Create plan
      tags:
      - admin
  /api/admin/plans/{plan_id}:
    delete:
      parameters:
      - description: plan uuid
        in: path
        name: plan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Delete plan
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: plan uuid
        in: path
        name: plan_id
        required: true
        type: string
      - description: updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdatePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Plan'
      security:
      - TelegramInitData: []
      summary: Update plan
      tags:
      - admin
  /api/admin/promo:
    get:
      parameters:
      - default: 50
        description: page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: List promo codes
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: promo params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePromoCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PromoCode'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Create promo code
      tags:
      - admin
  /api/admin/promo/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: params + count
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BulkPromoCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Bulk-create promo codes
      tags:
      - admin
  /api/admin/promo/deactivate:
    post:
      consumes:
      - application/json
      parameters:
      - description: '{code}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Deactivate promo code
      tags:
      - admin
  /api/admin/servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: List all servers (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: server config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Create server
      tags:
      - admin
  /api/admin/servers/{server_id}:
    delete:
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Delete server
      tags:
      - admin
    get:
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Get server (admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      - description: updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Update server
      tags:
      - admin
  /api/admin/servers/{server_id}/test:
    post:
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Test server XUI connection
      tags:
      - admin
  /api/admin/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: List settings
      tags:
      - admin
  /api/admin/settings/referral-bonus:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Get referral bonus percent
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: '{percent}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            format: float64
            type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Set referral bonus percent
      tags:
      - admin
  /api/admin/settings/referral-bonus-days:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Get referral bonus days
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: '{days}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Set referral bonus days
      tags:
      - admin
  /api/admin/settings/region-switch-price:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Get region switch price
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: '{price}'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SetRegionSwitchPriceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Set region switch price
      tags:
      - admin
  /api/admin/settings/topup-bonus:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Get topup bonus
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: '{percent}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            format: float64
            type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Set topup bonus
      tags:
      - admin
  /api/admin/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Admin dashboard stats
      tags:
      - admin
  /api/admin/users:
    get:
      parameters:
      - default: 50
        description: page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      - description: username/name search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListUsersResponse'
      security:
      - TelegramInitData: []
      summary: List users
      tags:
      - admin
  /api/admin/users/{user_id}:
    get:
      parameters:
      - description: user id (telegram)
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserWithSubscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Get user
      tags:
      - admin
  /api/admin/users/{user_id}/balance/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      - description: delta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AddBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Add to user balance
      tags:
      - admin
  /api/admin/users/{user_id}/balance/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      - description: new balance
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SetBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Set user balance
      tags:
      - admin
  /api/admin/users/{user_id}/ban:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      - description: reason + optional expiry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BanUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "409":
          description: already banned
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Ban user
      tags:
      - admin
  /api/admin/users/{user_id}/cash-payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      - description: plan + amount in RUB
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCashPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Create cash payment for user
      tags:
      - admin
  /api/admin/users/{user_id}/subscription/cancel:
    post:
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Cancel subscription
      tags:
      - admin
  /api/admin/users/{user_id}/subscription/extend:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      - description: days
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ExtendSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Extend subscription
      tags:
      - admin
  /api/admin/users/{user_id}/unban:
    post:
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Unban user
      tags:
      - admin
  /api/balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Current balance
      tags:
      - balance
...
schemes:
- https
securityDefinitions:
  TelegramInitData:
    description: |-
      Raw Telegram.WebApp.initData query string. HMAC-SHA256
      signed by the bot token; freshness window is 1 hour.
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"